requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
passlib[bcrypt]==1.7.4
pyjwt==2.10.1